Move ValidationError to DAGSpec from TaskSpec
